refac: rt_params_obj_info.c
